Implemented CNN
Implemented the CNN that will be used for the digit classification.
Built a small application that builds and loads the model from a given path if exists or else trains a model with MNIST training data and saves the model to the given path, then builds and tests the model with MNIST training data and plots a confusion matrix.
The model seems to have excellent performance