Built simple GUI for predicting self drawn digits
Built a simple GUI in which a user can draw a digit and trigger prediction. The user can trigger the training of the model via a button. After training the model will be saved for later use. After starting the program will search for a modell and will eventually request the user to start the trining if no model was found yet. The training process uses threading to prevent the GUI from freezing