Implemented viz skelton visualize.py
Implemented a skeleton for the visualization. visualize.py creates a second window for visualization. It can be loaded with a button added to the DigitClassifierApp. The new window has a button to start the visualization.

Currently the app clones the model from DigitClassifierApp, loads the current image via preprocess_image() and prints the activation of all conv2d and dense layers to the console for exploration.

For that some minor changes to model.py had to be done.

Additionally the format for the saving of the model was changed to.keras (models/cnn_model_full.keras).